% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Return a list of the first n numbers of the Fibonacci sequence.}
\usage{
fibonacci(n, first = 0, second = 1)
}
\arguments{
\item{n}{The number of Fibonacci numbers to return.}

\item{first}{The first number in the sequence.}

\item{second}{The second number in the sequence.}
}
\value{
A list of the first n numbers of the Fibonacci sequence.
}
\description{
Return a list of the first n numbers of the Fibonacci sequence.
}
\examples{
fibonacci(10)
fibonacci(10, 0, 0)
fibonacci(10, 1, 2)
}
