% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imc.R
\name{imc}
\alias{imc}
\title{Indice de masse corporel - fonction A}
\usage{
imc(masse, taille)
}
\arguments{
\item{masse}{numeric}

\item{taille}{numeric}
}
\value{
a single numeric value
}
\description{
Indice de masse corporel - fonction A
}
\examples{
imc(72, 1.82)

}
