% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imc2.R
\name{imc2}
\alias{imc2}
\title{Indice de masse corporel - fonction imc2()}
\usage{
imc2(masse, taille)
}
\arguments{
\item{masse}{numeric}

\item{taille}{numeric}
}
\value{
a single numeric value
}
\description{
Indice de masse corporel - fonction imc2()
}
\examples{
imc2(72, 1.82)

\dontrun{
imc2("75", "1.7")
}

}
