% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_num.R
\name{is_num}
\alias{is_num}
\title{is_num personnalise - function interne}
\usage{
is_num(x, y)
}
\arguments{
\item{x}{an integer , numeric}

\item{y}{an integer, numeric}
}
\value{

}
\description{
is_num personnalise - function interne
}
\examples{
is_num(25, "250")

}
\keyword{Internal}
