% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laurentspanu.R
\docType{package}
\name{laurentspanu}
\alias{laurentspanu}
\title{laurentspanu : package de deux fonctions}
\description{
Le package apporte deux fonctions:
imc et multi_import_excel
}
\section{Les fonctions}{

imc() et multi_import_excel
}

