% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_import_excel.R
\name{multi_import_excel}
\alias{multi_import_excel}
\title{Lire les onglets d'un fichier excel}
\usage{
multi_import_excel(file)
}
\arguments{
\item{file}{fichier xlsx en chaîne de caractères}
}
\value{
une liste
}
\description{
Lire les onglets d'un fichier excel
}
\examples{
multi_import_excel(file = system.file("extdata","datasets.xlsx",package = "laurentspanu"))
}
