% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tmp.R
\docType{data}
\name{tmp}
\alias{tmp}
\title{Documentation du data frame demo data/tmp.rda}
\format{
un dataframe avec 2 lignes et 5 variables:
\describe{
\item{Sepal.Length}{length du bidule}
\item{Sepal.Width}{width du bidule}
\item{Petal.Length}{length du truc}
\item{Petal.Length}{width du truc}
\item{Species}{espèce}
}
}
\usage{
tmp
}
\description{
Un jeu de data qui provient de jeu iris
il est constitué de 2 lignes et ne sert
à rien au fait, c est juste un essai.
}
\author{
Spanu \email{l.spanu@issep.be}
}
\keyword{data}
