% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_dflospa.R
\docType{data}
\name{df_lospa}
\alias{df_lospa}
\title{df_lospa df de test}
\format{
data.frame de 7 lignes et de 2 variables:
\describe{
\item{masse}{masse, en kg}
\item{taille}{taille, en m}
}
}
\usage{
df_lospa
}
\description{
Un df reprenant la masse et la taille de 7 individus. Utile pr tester fonction imc()
}
\author{
Spanu \email{l.spanu@issep.be}
}
\keyword{data}
