% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laurentspanu.R
\docType{package}
\name{laurentspanu}
\alias{laurentspanu}
\alias{laurentspanu-package}
\title{Package 'laurentspanu'}
\description{
Le package apporte deux fonctions:
imc et multi_import_excel
}
\section{Les fonctions}{

imc() et multi_import_excel
Ces fonction sont vraiment coolos
}

\author{
\strong{Maintainer}: Laurent Spanu \email{l.spanu@issep.be}

}
\keyword{internal}
